Jan 2025

+++ include/event_flag.h
@@ -146,7 +146,7 @@ void GetScriptVarRef(struct ScriptVarPtr *out, u8 *localVarBuf, s32 varId);
 s32 GetScriptVarValue(u8 *localVarBuf, s32 varId);
 s32 GetScriptVarArrayValue(u8 *localVarBuf, s32 varId, u16 idx);
 void SetScriptVarValue(u8 *localVarBuf, s32 varId, s32 val);
-void SetScriptVarArrayValue(u8 *localVarBuf, s32 varId, u16 idx, s32 val);
+void SetScriptVarArrayValue(u8 *localVarBuf, s32 varId, s32 idx, s32 val);
 s32 GetScriptVarArraySum(u8 *localVarBuf, s16 varId);
 void GetScriptVarScenario(s32 varId,u32 *outMain,u32 *outSub);
 void ScenarioCalc(s16 varId, s32 main, s32 sub);
